Try to avoid finding chains in different order
By only allowing certain expressions to be added first, e.g. if at the
current level we choose another expression, then we disallow
previously used expressions at this further down the chain.

This is much faster, alas it didn't find the 24 chain. I suspect it's
a bit of a gamble which expressions will be tried as the first 5 or 3,
and ruling out some of these paths simply misses some opportunities.

I'll remove it in the next commit, just wanted to have it committed.