Optimize order of expressions
This probably depends heavily on the input expressions, but for N=11 I
used counted the number of occurrences in the solutions, ordered them,
then ran this with the expressions in that order and reversed to see
the total chains looked at. This order affects how early some chains
bail out, and at least for this order (the same as the frequency,
smaller to bigger) the number was the smallest I've seen.