Reflect the positive path of a branch
In the majority of cases this check will be true, so have that as the
"default" body on top. It also saves a negation in theory.
In tests both versions performed quite similar, the compiler might
optimize the negation away anyway, but we might as well remove it from
the code.