Improve progress restoration and chunk termination
This is a much cleaner way to do it, and it's close to no overhead,
because the comparison to max_i has to happen anyway.

This actually made it a little slower. Looks like branching and
changing max_i makes the compiler generate slower code. I'll try to
address this in the next commit.