STRANGE: this useless code improves performance
Maybe it results in some paths that end up being faster than the code
without it. Might be worth investigating.

It might have to do with compare_choices_with_start_indices.
Perhaps that caches chain data just so?

However, in this version progress_check_done is true from the start,
so that path is never executed. It's weird.