Pick target functions if necessary in hungry algorithm
For the last w chain steps before the max length, where w is the
number of not-yet-seen target functions we don't have to apply
algorithm L at all, we KNOW each step must be a target function, and
any relevant target function available will do, we don't have to try
them all. So in this case don't apply algorithm L, which is expensive,
just generate the first level expressions and pick the first target
function among them.

This reduces the search space by 75%.