Fix chain extraction from function sets
Now it really generates the unique chains for each group, ignoring
order, but considering two chains different if all their functions are
the same but at least one function was derived with a different
instruction.

A group consists of all chains that result in the same 16-bit target
functions, i.e. the display for all digits looks the same for all
chains inside a group but different for any two chains from any two
different groups.

The chosen chain is canonical in the way that it'll pick any
instruction for earlier inputs/functions first and ignore it in other
chains following this branch.