Add configuration to check for shorter chains
The assumption is that we have good reason to assume MAX_LENGTH is the
minimal length of chains we'll find. So let the compiler optimize this
check away entirely, if we expect it to be pointless.

I've kept it in as a configuration, however, to document this, in case
there's ever unexpected behaviour, and to enable it again, if needed.