Define new bite sizes for hungry search
For lower N it turns out that the big bite sizes tend to be at the
beginning, and never above 30 (so far). Later on they usually are
quite low, so using the first 4 big bites as chunk inputs and small
bites later on this might be a manageable search space.