Give full-search fixed chunk size
This simplifies a lot of the logic around storing progress. It also
reduces the initial goto and the number of expected prefix numbers on
the command line to one constant: CHUNK_START_LENGTH, so there can't
be any accidents in getting this wrong.

The plan mode also now is hard coded to CHUNK_START_LENGTH, again
avoiding mismatches.

Finally, the PRINT_PROGRESS_LENGTH constant wasn't correctly used,
this fixes it.